Extract filter editor and can edit date
